requirements.txt
matplotlib==3.3.2
numpy==1.19.2
opencv-python==4.4.0.46
pandas==1.1.4
pillow==8.0.1
pytorch==1.7.0
scikit-learn==0.23.2
seaborn==0.11.0